backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
python-multipart>=0.0.20
httpx==0.28.1
pytesseract==0.3.13
Pillow>=10.4.0
sentence-transformers
faiss-cpu
requests
PyPDF2
google-genai
pypdf
torch
numpy
supabase